Board and physics changes
I added a rocket model to my board plus made a few other changes to its layout.

Additionally I shifted the layout of my board and the hysics of the project so that the ball goes down the board at the same rate as it would if it were falling in that direction.